feat: Implement the logic in WithdrawCommand
Request all necessary data from user (currency code, amount), check
their correctness.
Check whether there are enough funds in the account.
Withdraw money from the account using the new withdrawAmount() method.
Output to the user.
Sorting - from a larger denomination to a smaller one.
Display a message about a successful transaction.
Catch NotEnoughMoneyException.
In case of a shortage of money on the account, notify the user about it.

The logic of the main withdrawAmount method:
withdrawAmount method returns the minimum number of banknotes that the
requested amount is typed with.
If there are several options, then use the one with the maximum number
of banknotes of the highest denomination.
If we have found such an option for issuing money, then subtract all
these banknotes from the card in the manipulator.
withdrawAmount method throw NotEnoughMoneyException if it is impossible
to issue the requested amount in banknotes.